requirements.txt
configparser==3.5.0
Django==1.9.8
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
mysqlclient==1.3.7
